Don't repeat block name in endblock tag
Hi, not sure, if you agree and this would break anything for someone, 
but with this change the name of a block would not be used in the `endblock` tag.
My impression is, that this would be in line with the docs 
(https://docs.djangoproject.com/en/2.1/ref/templates/language/#templates) and for what
I have seen, it seems to be unusual to have the name there. When I am using the expansion,
I always have to remove the name again.